+++ llm_on_ray/common/load.py
@@ -2,12 +2,7 @@ import sys
 from typing import Any, Dict
 
 from llm_on_ray.common.logging import logger
-from llm_on_ray.common import dataset
-from llm_on_ray.common import tokenizer
-from llm_on_ray.common import model
-from llm_on_ray.common import optimizer
-from llm_on_ray.common import trainer
-from llm_on_ray.common import initializer
+from llm_on_ray.common import agentenv, dataset, initializer, model, optimizer, tokenizer, trainer
 
 
 def load_check_decorator(func):
@@ -114,6 +109,21 @@ def get_initializer(config: Dict[str, Any]):
     Factory = initializer.Initializer.registory.get(initializer_type)
     if Factory is None:
         raise ValueError(f"there is no {initializer_type} initializer.")
+    try:
+        _ = Factory(config)
+    except Exception as e:
+        logger.critical(f"{Factory.__name__} init error: {e}", exc_info=True)
+        exit(1)
+    return _
+
+
+@load_check_decorator  # type: ignore # noqa: F405 # may be undefined, or defined from star imports
+def get_agentenv(config: Dict[str, Any]):
+    logger.info(f"{sys._getframe().f_code.co_name} config: {config}")
+    agentenv_type = config.get("type", None)
+    Factory = agentenv.AgentEnv.registory.get(agentenv_type)
+    if Factory is None:
+        raise ValueError(f"there is no {agentenv_type} AgentEnv.")
     try:
         _ = Factory(config)
     except Exception as e:
